+++ src/ConfigurationData.php
@@ -3,17 +3,22 @@
 declare(strict_types=1);
 
 namespace Neucore\Plugin\Mumble;
+
 class ConfigurationData
 {
     public function __construct(
+        public string $nickname,
+
         public array $groupsToTags,
 
-        public ?int $bannedGroup,
+        public bool $mainTagReplacesCorporationTicker,
 
         /**
          * @var array<array<string>>
          */
         public array $additionalTagGroups,
+
+        public ?int $bannedGroup,
     ) {
     }
 }