+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 - Breaking: Moved NEUCORE_MUMBLE_PLUGIN_BANNED_GROUP environment variable to the configuration data.
 - Breaking: Moved the hard coded additional tags (CEO, pronouns) to the configuration data. The example configuration
   in plugin.yml contains a compatible configuration.
+- Breaking: Added "Nickname" configuration value.
+- Breaking: Added "MainTagReplacesCorporationTicker" configuration value.
 - Change: Previously the last found pronoun from the list in the configuration data was used, now it's the first.
 - NEUCORE_MUMBLE_PLUGIN_DB_USERNAME and NEUCORE_MUMBLE_PLUGIN_DB_PASSWORD are now optional: user and password
   can be added to NEUCORE_MUMBLE_PLUGIN_DB_DSN.